// FE/CV_app_user/src/modules/Home/shared/enums.ts
export enum FORM_DATA_FIELD {
  name = 'name',
  userName = 'userName',
  // first_name = 'first_name',
  // full_name = 'full_name',
  phone = 'phone',
  gender_id = 'gender_id',
  address = 'address',
  probation_period = 'probation_period',
  enterprise_name = 'enterprise_name',
  city_id = 'city_id',
  district_id = 'district_id',
  email = 'email',
  password = 'password',
  new_password = 'new_password',
  remember_me = 'remember_me',
  confirmPassword = 'confirmPassword',
  receive_news = 'receive_news',
  level = 'level',
  field = 'field',
  verify_code = 'code',
  avatar = 'avatar',
  oldPass = 'oldPass',
}

export enum ENTERPRISE_FORM {
  id = 'id',
  name = 'name',
  abbreviation_name = 'abbreviation_name',
  address = 'address',
  scale_id = 'scale_id',
  city_id = 'city_id',
  district_id = 'district_id',
  ward_id = 'ward_id',
  map_url = 'map_url',
  phone = 'phone',
  career_field_id = 'career_field_id',
  website_url = 'website_url',
  introduce = 'introduce',
  user_id = 'user_id',
  avatar = 'avatar',
  represent = 'represent',
  plan = 'plan',
  phone_verified = 'phone_verified',
  enterprise_name = 'enterprise_name',
}

// enum getopt type
export enum GET_OTP_TYPE {
  SIGNUP = 1,
  CHANGE_ENTERPRISE_PHONE = 2,
  CHANGE_REPRESENT_PHONE = 3,
}
export enum RECRUITMENT_DATA_FIELD {
  title = 'title',
  city_id = 'city_id',
  district_id = 'district_id',
  map_url = 'map_url',
  avatar = 'avatar',
  form_of_work_id = 'form_of_work_id',
  diploma_id = 'diploma_id',
  experience_id = 'experience_id',
  level_id = 'level_id',
  gender_id = 'gender_id',
  career_field_id = 'career_field_id',
  career_id = 'career_id',
  deadline = 'deadline',
  probationary_period_id = 'probationary_period_id',
  salary_type_id = 'salary_type_id',
  salary_min = 'salary_min',
  salary_max = 'salary_max',
  tags = 'tags',
  overview = 'overview',
  requirement = 'requirement',
  benefit = 'benefit',
  contact_name = 'contact_name',
  contact_phone = 'contact_phone',
  contact_email = 'contact_email',
  status_id = 'status_id',
  // caching = 'caching',
  image_url = 'image_url',
  address = 'address',
}
export enum RECRUITMENT_UPDATE_DATA_FIELD {
  title = 'title',
  city = 'city',
  district = 'district',
  map_url = 'map_url',
  avatar = 'avatar',
  form_of_work = 'form_of_work',
  diploma = 'diploma',
  experience = 'experience',
  level = 'level',
  gender = 'gender',
  career_field_id = 'career_field_id',
  career_id = 'career_id',
  deadline = 'deadline',
  probationary_period = 'probationary_period',
  salary_type = 'salary_type',
  salary_min = 'salary_min',
  salary_max = 'salary_max',
  tags = 'tags',
  overview = 'overview',
  requirement = 'requirement',
  benefit = 'benefit',
  contact_name = 'contact_name',
  contact_phone = 'contact_phone',
  contact_email = 'contact_email',
  status = 'status_id',
  // caching = 'caching',
  image_url = 'image_url',
  address = 'address',
}